requirements.txt
Pillow
fastapi
jinja2
uvicorn
python-multipart